Cleaned up Linux implementation
