+++ system_idle_linux/.gitignore
@@ -9,3 +9,4 @@ pubspec.lock
 # Contains compiled native code
 dist/
 *.so
+*.o
